SchrodingerToolkit

Documentation for SchrodingerToolkit.

SchrodingerToolkit.AbstractRuntimeStats

SchrodingerToolkit.OffsetAllZero

SchrodingerToolkit.OffsetNonZero

SchrodingerToolkit.OffsetUniqueZero

SchrodingerToolkit.SchrodingerPDEComponent

SchrodingerToolkit.SchrodingerPDENonPolynomic

SchrodingerToolkit.SchrodingerPDEPolynomic

Base.eltype

Base.eltype

Base.ndims

Base.ndims

SchrodingerToolkit.get_boundary

SchrodingerToolkit.get_boundary

SchrodingerToolkit.get_component

SchrodingerToolkit.get_component

SchrodingerToolkit.get_f

SchrodingerToolkit.get_f

SchrodingerToolkit.get_field

SchrodingerToolkit.get_field

SchrodingerToolkit.get_time_boundary

SchrodingerToolkit.get_time_boundary

SchrodingerToolkit.get_σ

SchrodingerToolkit.get_σ

SchrodingerToolkit.get_ψ

SchrodingerToolkit.get_ψ

SchrodingerToolkit.matrix_to_vector

SchrodingerToolkit.test_if_zero

SchrodingerToolkit.vcanonical

SchrodingerToolkit.vfirst

SchrodingerToolkit.vlast

SchrodingerToolkit.vones

SchrodingerToolkit.vseq

SchrodingerToolkit.vundef

SchrodingerToolkit.vzeros

SchrodingerToolkit.AbstractRuntimeStats — Type

AbstractRuntimeStats{IntType,FloatType,ArrayType}

This abstract type provides basic tracking for runtime of PDE simulation.

Functions that you need to implement your own Stats type:

length(Stats): Current number of taken samples.

islocked(Stats): Returns true if the Statsis locked meaning that you can not move from samplenton+1`. However some fields can be updated depending on your design.

initialize_stats(Stats,ncomponents,log_freq,τ,timesteps,log_solverinfo): Here you should initialize the Stats type. ncomponents is the number of components in the PDE, log_freq is the frequency of logging, τ is the time step, timesteps is the total number of time steps, and log_solverinfo is a boolean that indicates whether to log solver information.

update_system_energy!(Stats,energy,index): Update the system energy at time step index.

update_power!(Stats,power,index): Update the power at time step index.

advance!(Stats): Move from sample n to n+1.

serialize(Stats,path): Serialize the Stats type to a file.

deserialize(Stats,path): Deserialize the Stats type from a file.

Optional functions:

update_stats!(Stats,time,PDE,Grid,Mem,ItStop): For iterative solvers

update_stats!(Stats,time,PDE,Grid,Mem): For direct solvers

Examples

SchrodingerToolkit.OffsetAllZero — Type

No dimension has the zero offset

SchrodingerToolkit.OffsetNonZero — Type

Every dimension has the zero offset

SchrodingerToolkit.OffsetUniqueZero — Type

Only the first dimension has the zero offset

SchrodingerToolkit.SchrodingerPDEComponent — Type

PDE component

SchrodingerToolkit.SchrodingerPDENonPolynomic — Type

Generic handler for non polynomic potentials

SchrodingerToolkit.SchrodingerPDEPolynomic — Type

Optimized structure for polynomical potentials

Base.eltype — Method

Type of the elements for the Schrodinger PDE

Base.eltype — Method

Type of the elements for the Schrodinger PDE

Base.ndims — Method

Number of dimensions for the Schrodinger PDE

Base.ndims — Method

Number of dimensions for the Schrodinger PDE

SchrodingerToolkit.get_boundary — Method

Boundary start and end point at index dimension

SchrodingerToolkit.get_boundary — Method

Boundary start and end point at index dimension

SchrodingerToolkit.get_component — Method

Obtain the component of the Schrodinger PDE at index

SchrodingerToolkit.get_component — Method

Obtain the component of the Schrodinger PDE at index

SchrodingerToolkit.get_f — Method

Function for the Schrodinger PDE component

SchrodingerToolkit.get_f — Method

Function for the Schrodinger PDE component

SchrodingerToolkit.get_field — Method

Potential for the Schrodinger PDE

SchrodingerToolkit.get_field — Method

Potential for the Schrodinger PDE

SchrodingerToolkit.get_time_boundary — Method

Finish time for the Schrodinger PDE

SchrodingerToolkit.get_time_boundary — Method

Finish time for the Schrodinger PDE

SchrodingerToolkit.get_σ — Method

Dispersion coefficient for the Schrodinger PDE component at index

SchrodingerToolkit.get_σ — Method

Dispersion coefficient for the Schrodinger PDE component at index

SchrodingerToolkit.get_ψ — Method

Initial condition for the Schrodinger PDE component

SchrodingerToolkit.get_ψ — Method

Initial condition for the Schrodinger PDE component

SchrodingerToolkit.matrix_to_vector — Method

S = matrix_to_vector(M)

Return the dense vector storage type S related to the dense matrix storage type M.

SchrodingerToolkit.test_if_zero — Function

v = test_if_zero(V,idx)

Return true if the element at position idx of the vector V is zero.

SchrodingerToolkit.vcanonical — Function

v = vcanonical(S, n, idx)

Create a vector of storage type S of length n where the position idx is one and any other position is zero.

SchrodingerToolkit.vfirst — Function

v = vfirst(V)

Return the first element of the vector V.

SchrodingerToolkit.vlast — Function

v = vlast(V)

Return the last element of the vector V.

SchrodingerToolkit.vones — Method

v = vones(S, n)

Create a vector of storage type S of length n only composed of one.

SchrodingerToolkit.vseq — Method

v = vseq(S, n)

Create a vector of storage type S of length n where the position i is i.

SchrodingerToolkit.vundef — Method

v = vundef(S, n)

Create an uninitialized vector of storage type S of length n.

SchrodingerToolkit.vzeros — Method

v = vzeros(S, n)

Create a vector of storage type S of length n only composed of zero.