SchrodingerToolkit

Documentation for SchrodingerToolkit.

SchrodingerToolkit.AbstractBackend

SchrodingerToolkit.AbstractRuntimeStats

SchrodingerToolkit.AbstractSolverMethod

SchrodingerToolkit.CPUBackend

SchrodingerToolkit.GPUBackend

SchrodingerToolkit.OffsetAllZero

SchrodingerToolkit.OffsetNonZero

SchrodingerToolkit.OffsetUniqueZero

SchrodingerToolkit.SchrodingerPDE

SchrodingerToolkit.SchrodingerPDEComponent

SchrodingerToolkit.SchrodingerPDENonPolynomic

SchrodingerToolkit.SchrodingerPDEPolynomic

SchrodingerToolkit.xPUBackend

Base.eltype

Base.eltype

Base.ndims

Base.ndims

SchrodingerToolkit.calculate_diff_system_energy

SchrodingerToolkit.calculate_diff_system_power

SchrodingerToolkit.estimate_timesteps

SchrodingerToolkit.evaluate_ψ

SchrodingerToolkit.evaluate_ψ

SchrodingerToolkit.evaluate_ψ!

SchrodingerToolkit.get_V

SchrodingerToolkit.get_V

SchrodingerToolkit.get_boundary

SchrodingerToolkit.get_boundary

SchrodingerToolkit.get_component

SchrodingerToolkit.get_component

SchrodingerToolkit.get_f

SchrodingerToolkit.get_f

SchrodingerToolkit.get_field

SchrodingerToolkit.get_field

SchrodingerToolkit.get_optimized

SchrodingerToolkit.get_optimized

SchrodingerToolkit.get_time_boundary

SchrodingerToolkit.get_time_boundary

SchrodingerToolkit.get_Γ

SchrodingerToolkit.get_Γ

SchrodingerToolkit.get_σ

SchrodingerToolkit.get_σ

SchrodingerToolkit.get_ψ

SchrodingerToolkit.get_ψ

SchrodingerToolkit.iscpu

SchrodingerToolkit.isgpu

SchrodingerToolkit.isxpu

SchrodingerToolkit.junction!

SchrodingerToolkit.matrix_to_vector

SchrodingerToolkit.measure

SchrodingerToolkit.ncomponents

SchrodingerToolkit.step!

SchrodingerToolkit.test_if_zero

SchrodingerToolkit.update_component!

SchrodingerToolkit.update_solver_info!

SchrodingerToolkit.update_stats!

SchrodingerToolkit.vcanonical

SchrodingerToolkit.vfirst

SchrodingerToolkit.vlast

SchrodingerToolkit.vones

SchrodingerToolkit.vseq

SchrodingerToolkit.vundef

SchrodingerToolkit.vzeros

SchrodingerToolkit.AbstractBackend — Type

Abstract type for our available backends

SchrodingerToolkit.AbstractRuntimeStats — Type

AbstractRuntimeStats{IntType,FloatType,ArrayType}

This abstract type provides basic tracking for runtime analysis of PDE simulation.

Functions that you need to implement your own Stats type:

length(Stats): Current number of taken samples.

islog(Stats): Returns true if the Stats is logging linear solver data.

islocked(Stats): Returns true if the Stats is locked meaning that you can not modify it. This is used to prevent data corruption after deserialization.

initialize_stats(::Type{FloatVectorType}, ::Type{IntVectorType},ncomponents,log_freq,time_steps,islog_solver_info::Bool): Initialize the memory underlying Stats type.

startup_stats!(Stats,start_power,start_energy): Initialize the power and energy at time step 0. This is usally stored at the last index of the memory.

update_power!(Stats,power,index): Update the power at time step index.

update_energy!(Stats,energy,index): Update the energy at time step index.

advance_iteration!(Stats): Move from sample n to n+1.

serialize(Stats,path): Serialize the Stats type to a file.

deserialize(::Type{Stats},path): Deserialize the Stats type from a file. The type here is for dispatching purposes.

system_power(Stats,component_idx,index): Get the power at component_idx in time step index.

system_power(Stats,component_idx): Get the full internal power vector at component_idx.

system_energy(Stats,index): Get the energy at time step index.

system_energy(Stats): Get the full internal energy vector.

start_power(Stats): Get the power for all components at time step 0.

start_power(Stats,component_idx): Get the power at component_idx at time step 0.

start_energy(Stats): Get the energy at time step 0.

get_solver_time(Stats,index): Get the solver time at time step index.

get_solver_iterations(Stats,index): Get the solver iterations at time step index.

Note 1: Any attempt to modify the Stats type after it is locked must throw an error or a warning.

Note 2: In general is preferred to use CPU vectors for Stats type and taking care of the conversion to other types when needed. This is because indexing is slow and disallowed in GPU arrays.

Note 3: As an API decision, to reduce boilerplate code is expected that when you intialize any method, then this structure is initialized too. This is because the Stats type is used to store incoming data from the simulation and data types are known at that moment.

Optional functions:

update_stats!(Stats,time,PDE,Grid,Mem,ItStop): For iterative solvers

update_stats!(Stats,time,PDE,Grid,Mem): For direct solvers

Examples

SchrodingerToolkit.AbstractSolverMethod — Type

AbstractSolverMethod{RealType}

This is the abstract type for all solver methods. It is parametrized by the type of the real numbers used in the solver.

# How to define a new solver method
To define a new solver method, you need to define a new struct that is a subtype of this abstract type. The struct must have the following fields:
- `Grid`: The grid on which the solver will operate.

# Expected initialization
The solver method must be initialized with the following arguments:
- `config`: The basic configuration of the solver.
- `grid`: The grid on which the solver will operate.

# Required methods
The solver method must implement the following methods:
- `step!`: Perform a single step of the solver.

# Methods related to statistics
[TODO]

SchrodingerToolkit.CPUBackend — Type

This represents a CPU based backend

SchrodingerToolkit.GPUBackend — Type

This represents a GPU based backend

SchrodingerToolkit.OffsetAllZero — Type

No dimension has the zero offset

SchrodingerToolkit.OffsetNonZero — Type

Every dimension has the zero offset

SchrodingerToolkit.OffsetUniqueZero — Type

Only the first dimension has the zero offset

SchrodingerToolkit.SchrodingerPDE — Type

SchrodingerPDE

This is an abstract type that represents a Schrodinger PDE. This type is used to define a Schrodinger PDE structure.

Parameters:

- `N::Int`: Number of dimensions
- `RealType`: Type for Real numbers. This is used to define the internal algorithms and data structures.

Is important to note, that the derived structures must handle boundary conditions, initial conditions and important parts of the PDE.

SchrodingerToolkit.SchrodingerPDEComponent — Type

SchrodingerPDEComponent

This structure represents a component of the Schrodinger PDE. This is a simple structure that contains the necessary information
to solve the Schrodinger PDE for a single component.

Fields:

- `σ::Tv`: Dispersion coefficient => Must be non negative
- `f::Fn`: ∂∇F/∂xᵢ(x) = f(x)
- `ψ::InitialCondition`: Initial condition 
- `V::TrappingPotential`: External trapping potential
- `Γ::Γv`: Josephson Junction Coefficient => In some documents this coefficient is the same for all components but I guess it can be different under some circumstances.

By default `V` and `Γ` are set to `nothing` but they can be set to a trapping potential and a Josephson junction coefficient respectively if you want to work with a Gross Pitaevsky PDE.

SchrodingerToolkit.SchrodingerPDENonPolynomic — Type

SchrodingerPDENonPolynomic

This structure represents a Schrodinger PDE with non-polynomical potentials.

Fields:

- `boundaries::NTuple{N,NTuple{2,Tv}}`: Boundaries of the cartesian domain where you want to solve the PDE.
- `components::MComp`: Components of the Schrodinger PDE
- `F::Potential`: Non-polynomical potential
- `T::Tv`: Final time

SchrodingerToolkit.SchrodingerPDEPolynomic — Type

SchrodingerPDEPolynomic

This structure represents a Schrodinger PDE with polynomical potentials.

Fields:

- `boundaries::NTuple{N,NTuple{2,Tv}}`: Boundaries of the cartesian domain where you want to solve the PDE.
- `components::MComp`: Components of the Schrodinger PDE
- `F::Potential`: Polynomical potential
- `N::Optimized`: Optimized structure for polynomical potentials. Here is stored an auxiliary function without the catastrophic cancellation problem.
- `T::Tv`: Final time

SchrodingerToolkit.xPUBackend — Type

This represents a heterogeneous based backend (Can use a mixture of CPU and GPU)

Base.eltype — Method

Type of the elements for the Schrodinger PDE

Base.eltype — Method

Type of the elements for the Schrodinger PDE

Base.ndims — Method

Number of dimensions for the Schrodinger PDE

Base.ndims — Method

Number of dimensions for the Schrodinger PDE

SchrodingerToolkit.calculate_diff_system_energy — Method

calculate_diff_system_energy(stats)

Calculate the absolute value of the difference between the system energy and the energy at time step `0`.

# Arguments
- `stats::Stats`: The stats structure.

SchrodingerToolkit.calculate_diff_system_power — Method

calculate_diff_system_power(stats)

Calculate the absolute value of the difference between the system power and the power at time step `0`.

# Arguments
- `stats::Stats`: The stats structure.

SchrodingerToolkit.estimate_timesteps — Method

Estimate the number of timesteps for the Schrodinger PDE

SchrodingerToolkit.evaluate_ψ! — Method

Evaluation of the initial condition for the Schrodinger PDE where the output is stored in Container

SchrodingerToolkit.evaluate_ψ — Method

Evaluation of the initial condition for the Schrodinger PDE

SchrodingerToolkit.evaluate_ψ — Method

Evaluation of the initial condition for the Schrodinger PDE

SchrodingerToolkit.get_V — Method

Trapping potential for the i-th Schrodinger PDE component

SchrodingerToolkit.get_V — Method

Trapping potential for the i-th Schrodinger PDE component

SchrodingerToolkit.get_boundary — Method

Boundary start and end point at index dimension

SchrodingerToolkit.get_boundary — Method

Boundary start and end point at index dimension

SchrodingerToolkit.get_component — Method

Obtain the component of the Schrodinger PDE at index

SchrodingerToolkit.get_component — Method

Obtain the component of the Schrodinger PDE at index

SchrodingerToolkit.get_f — Method

Function for the Schrodinger PDE component

SchrodingerToolkit.get_f — Method

Function for the Schrodinger PDE component

SchrodingerToolkit.get_field — Method

Potential for the Schrodinger PDE

SchrodingerToolkit.get_field — Method

Potential for the Schrodinger PDE

SchrodingerToolkit.get_optimized — Method

Non polynomial potentials can't be optimized so return nothing

SchrodingerToolkit.get_optimized — Method

Simplified polynomial to evade catastrophic cancellation

SchrodingerToolkit.get_time_boundary — Method

Finish time for the Schrodinger PDE

SchrodingerToolkit.get_time_boundary — Method

Finish time for the Schrodinger PDE

SchrodingerToolkit.get_Γ — Method

Josephson Junction Coefficient for the i-th Schrodinger PDE component

SchrodingerToolkit.get_Γ — Method

Josephson Junction Coefficient for the i-th Schrodinger PDE component

SchrodingerToolkit.get_σ — Method

Dispersion coefficient for the Schrodinger PDE component at index

SchrodingerToolkit.get_σ — Method

Dispersion coefficient for the Schrodinger PDE component at index

SchrodingerToolkit.get_ψ — Method

Initial condition for the Schrodinger PDE component

SchrodingerToolkit.get_ψ — Method

Initial condition for the Schrodinger PDE component

SchrodingerToolkit.iscpu — Function

Function to check if a backend is a CPU based backend

SchrodingerToolkit.isgpu — Function

Function to check if a backend is a GPU based backend

SchrodingerToolkit.isxpu — Function

Function to check if a backend is a heterogeneous based backend

SchrodingerToolkit.junction! — Method

junction!(SPDE::PDEeq, memory, output)

Check if the Schrodinger PDE has a Josephson Junction and return the sum of the current state if it has one.

SchrodingerToolkit.matrix_to_vector — Method

S = matrix_to_vector(M)

Return the dense vector storage type S related to the dense matrix storage type M.

SchrodingerToolkit.measure — Method

measure(A::PeriodicGrid{V,T,R,N}) where {V<:Integer,T<:Real,R<:AbstractRange{T},N}

Return the measure of the grid.

SchrodingerToolkit.ncomponents — Method

Number of components for the Schrodinger PDE

SchrodingerToolkit.step! — Function

With this function you can advance one step in time.

    The implementation of this function is mandatory.

    From the user perspective, this function should be called to advance the solution one step in time.

    The expected parameters are:
        - `method`: The method to be used to advance the solution.
        - `memory`: The memory of the method.
            Expectations about memory are: 
                - The memory must be allocated by the method at initialization.
                - The memory must be updated by the method.
                - You must provide a method called `current_state!(memory)` that allows the caller to a reference to the current state (time step).
                - You must provide a method called `current_state(memory)` that allows the caller to get a copy of the current state of (time step).
        - `stats`: The statistics of the method.
                - If you want to log statistics and you perform some kind of component wise operation then:
                    - You must call `update_solver_info!(stats, time, niter)` to store every time spent at linear solvers.
                - In general 
        - `PDE`: The PDE to be solved.
        - `config`: The configuration of the solver.
    
    Here dispatch is used to call the correct version of step!.

    The return of this function must be the time taken to advance one step in time.

SchrodingerToolkit.test_if_zero — Function

v = test_if_zero(V,idx)

Return true if the element at position idx of the vector V is zero.

SchrodingerToolkit.update_component! — Function

With this function you can update a component of the problem.

    The implementation of this function is suggested but not mandatory.

SchrodingerToolkit.update_solver_info! — Method

update_solver_info!(stats, time, iterations)

Update the linear solver information in the stats structure.

# Arguments
- `stats::Stats`: The stats structure.
- `time::FloatType`: The time spent in the solver.
- `iterations::IntType`: The number of iterations performed in the solver.

SchrodingerToolkit.update_stats! — Method

update_stats!(stats, step_time, power_per_component, sys_energy)

Update the stats structure with the current time, power per component and system energy.

# Arguments
- `stats::Stats`: The stats structure.
- `step_time::FloatType`: Amount of time spent in the current step.
- `power_per_component`: The power per component.
- `sys_energy::FloatType`: The system energy.

SchrodingerToolkit.vcanonical — Function

v = vcanonical(S, n, idx)

Create a vector of storage type S of length n where the position idx is one and any other position is zero.

SchrodingerToolkit.vfirst — Function

v = vfirst(V)

Return the first element of the vector V.

SchrodingerToolkit.vlast — Function

v = vlast(V)

Return the last element of the vector V.

SchrodingerToolkit.vones — Method

v = vones(S, n)

Create a vector of storage type S of length n only composed of one.

SchrodingerToolkit.vseq — Method

v = vseq(S, n)

Create a vector of storage type S of length n where the position i is i.

SchrodingerToolkit.vundef — Method

v = vundef(S, n)

Create an uninitialized vector of storage type S of length n.

SchrodingerToolkit.vzeros — Method

v = vzeros(S, n)

Create a vector of storage type S of length n only composed of zero.